Add utilities class and functionlity to invert the dependency graph
